Only create new symlink if outputFileName does not exist

+++ src/utils/processes/utils.nf
@@ -447,7 +447,9 @@ process SC__PUBLISH_PROXY {
             null
         )
         """
-        ln -s $f "${outputFileName}"
+        if [ ! -f ${outputFileName} ]; then
+            ln -s $f "${outputFileName}"
+        fi
         """
 
 }
